UPDATE: Primes #10 Refactoring
in isPrime, only mod by primes, to avoid non necessary composite numbers
it saved only 3 seconds, a bit surprised by such small improvement. it
is possibly because of the array indexing slowed it down. if this is
case, it is not a good choice. 10K is not a big size for an array. it
may come better if array gets bigger,