UPDATE: add IncrementalSieve for Primes calculation
This was developed very eariler in the week, while I had not time to do
proper exercise, but played a bit with this algorithm.
the is a modified version of the Sieve of Eratosthenes.

while I was doing this, I found the php array indexing issues with very
large numbers, so it has to be assigned using a string casing.
my version omitted all the even numbers, and add primes only once in the
running composites array. which should have increased the performance a
bit.